requirements.txt
# Core dependencies
langchain>=0.0.267
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
chromadb>=0.4.6
Flask>=2.3.2
huggingface-hub>=0.16.4
transformers>=4.32.0
torch>=2.2.1
arxiv>=1.4.7
scholarly>=1.7.11
beautifulsoup4>=4.12.2
requests>=2.31.0
pypdf2>=3.0.1
gradio>=3.40.0
accelerate>=0.23.0
bitsandbytes>=0.41.0
streamlit>=1.43.2
python-dotenv==1.0.1

fastapi==0.110.0
uvicorn==0.30.0
# sentence-transformers==2.5.1

# For deployment
gunicorn>=21.2.0



# # API
# fastapi==0.110.0
# uvicorn==0.30.0
# pydantic==2.6.4
# python-multipart==0.0.9

# # Data Retrieval
# arxiv==2.1.0
# scholarly==1.7.11
# beautifulsoup4==4.12.3
# requests==2.31.0
# pypdf2==3.0.1

# # Indexing
# sentence-transformers==2.5.1
# faiss-cpu==1.7.4

# # Summarization
# transformers==4.42.3
# torch==2.2.1
# langchain==0.1.13
# accelerate==0.31.0

# # UI
# streamlit==1.36.0

# # Utilities
# python-dotenv==1.0.1

